crawler/requirements.txt
Scrapy==2.11.2
PyYAML==6.0.2
